# perf(webidl/DOMString): don't wrap string primitives